requirements.txt
hyperliquid-python-sdk>=0.5.0
python-dotenv>=1.0.0
pandas>=2.0.0
eth-account>=0.10.0